third_party/WebKit/LayoutTests/platform/chromium-linux/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-decoration-line-expected.txt

Test to make sure text-decoration-line property returns values properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Initial value:
PASS e.style.getPropertyCSSValue('text-decoration-line') is null
PASS computedStyle.textDecorationLine is 'none'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'none'

Initial value (explicit):
PASS e.style.textDecorationLine is 'initial'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValue]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'initial'
PASS computedStyle.textDecorationLine is 'none'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'none'

Value 'none':
PASS e.style.textDecorationLine is 'none'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSPrimitiveValue]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'none'
PASS computedStyle.textDecorationLine is 'none'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'none'

Value 'underline':
PASS e.style.textDecorationLine is 'underline'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'underline'
PASS computedStyle.textDecorationLine is 'underline'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'underline'

Value 'overline':
PASS e.style.textDecorationLine is 'overline'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'overline'
PASS computedStyle.textDecorationLine is 'overline'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'overline'

Value 'line-through':
PASS e.style.textDecorationLine is 'line-through'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'line-through'
PASS computedStyle.textDecorationLine is 'line-through'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'line-through'

Value 'underline overline line-through':
PASS e.style.textDecorationLine is 'underline overline line-through'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'underline overline line-through'
PASS computedStyle.textDecorationLine is 'underline overline line-through'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'underline overline line-through'

Value 'blink' (invalid, last valid value is used):
PASS e.style.textDecorationLine is 'underline overline line-through'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'underline overline line-through'
PASS computedStyle.textDecorationLine is 'underline overline line-through'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'underline overline line-through'

Value '':
PASS e.style.getPropertyCSSValue('text-decoration-line') is null
PASS computedStyle.textDecorationLine is 'none'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSPrimitiveValue]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'none'

Parent gets 'underline' value:
PASS e.style.textDecorationLine is 'underline'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'underline'
PASS computedStyle.textDecorationLine is 'underline'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'underline'

Ancestor should explicitly inherit value from parent when 'inherit' value is used:
PASS e.style.textDecorationLine is 'inherit'
PASS e.style.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValue]'
PASS e.style.getPropertyCSSValue('text-decoration-line').cssText is 'inherit'
PASS computedStyle.textDecorationLine is 'underline'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('text-decoration-line').cssText is 'underline'

PASS successfullyParsed is true

TEST COMPLETE
